Move counterexample beautification to goto_checker
This is language-independent and can be used by
incremental goto checkers that provide traces.
