+++ src/cbmc/Makefile
@@ -4,7 +4,6 @@ SRC = all_properties.cpp \
       cbmc_languages.cpp \
       cbmc_main.cpp \
       cbmc_parse_options.cpp \
-      counterexample_beautification.cpp \
       fault_localization.cpp \
       xml_interface.cpp \
       # Empty last line
