+++ src/goto-checker/counterexample_beautification.cpp
