+++ src/goto-checker/counterexample_beautification.h
@@ -9,8 +9,8 @@ Author: Daniel Kroening, kroening@kroening.com
 /// \file
 /// Counterexample Beautification
 
-#ifndef CPROVER_CBMC_COUNTEREXAMPLE_BEAUTIFICATION_H
-#define CPROVER_CBMC_COUNTEREXAMPLE_BEAUTIFICATION_H
+#ifndef CPROVER_GOTO_CHECKER_COUNTEREXAMPLE_BEAUTIFICATION_H
+#define CPROVER_GOTO_CHECKER_COUNTEREXAMPLE_BEAUTIFICATION_H
 
 #include <util/namespace.h>
 
@@ -46,4 +46,4 @@ protected:
   symex_target_equationt::SSA_stepst::const_iterator failed;
 };
 
-#endif // CPROVER_CBMC_COUNTEREXAMPLE_BEAUTIFICATION_H
+#endif // CPROVER_GOTO_CHECKER_COUNTEREXAMPLE_BEAUTIFICATION_H
